Renamed inspect-plugins.py to plugins.py, updated to use argparse and added binary utilities
